pyproject.toml
[project]
name = "avito-cup-ml-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "pandas>=2.2.3",
    "polars>=1.29.0",
    "pyarrow>=20.0.0",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "setuptools>=80.7.1",
]

[dependency-groups]
baseline = [
    "catboost>=1.2.8",
    "implicit>=0.7.2",
    "lightgbm>=4.6.0",
]
dev = [
    "ipykernel>=6.29.5",
    "mypy>=1.15.0",
    "ruff>=0.11.10",
]
model-experiments = [
    "clearml>=1.18.0",
    "seaborn>=0.13.2",
    "shap>=0.47.2",
]
nn-models = [
    "pytorch-lightning>=2.5.1.post0",
    "torch>=2.7.0",
    "transformers>=4.52.3",
]
